+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 6.6.52
 - Addressbook: Added 'load more' butten to SelectDialogPanel
+- History: Custom fields are logged now too.
+- Smime: fixed upload certificate with error 'Certificate already exists'
 
 11-03-2022 6.6.51
 - Projects: Added state to all collapsible panels